refactor: remove 'server' imports
