+++ bciers/apps/dashboard/app/onboarding/page.tsx
@@ -5,7 +5,7 @@ import Button from "@mui/material/Button";
 import Grid from "@mui/material/Grid/Grid";
 import events from "@/dashboard/app/data/home/events.json";
 import { signIn } from "next-auth/react";
-import { getEnvValue } from "@bciers/utils/actions";
+import { getEnvValue } from "@bciers/actions";
 import {
   bcObpsLink,
   bcObpsGuidanceLink,
