+++ bciers/apps/registration/app/components/facilities/fetchFacilitiesPageData.ts
@@ -1,5 +1,5 @@
 import buildQueryParams from "@bciers/utils/buildQueryParams";
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { FacilitiesSearchParams } from "./types";
 
 // ↓ fetch data from server
