+++ bciers/apps/registration/middlewares/withAuthorizationRegistration.tsx
@@ -4,9 +4,9 @@ import {
   NextRequest,
   NextResponse,
 } from "next/server";
-import { MiddlewareFactory } from "@bciers/middlewares/server";
+import { MiddlewareFactory } from "@bciers/middlewares";
 import { IDP } from "@bciers/utils/enums";
-import { getToken } from "@bciers/actions/server";
+import { getToken } from "@bciers/actions";
 
 /*
 Access control logic is managed using Next.js middleware and NextAuth.js authentication JWT token.
