+++ bciers/apps/registration1/app/(authenticated)/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import Tiles from "@/app/components/navigation/Tiles";
 import { getUserOperator } from "@/app/components/routes/select-operator/Page";
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { FrontEndRoles } from "@bciers/utils/enums";
 import Card from "@mui/material/Card";
 import Typography from "@mui/material/Typography";
