+++ bciers/apps/registration1/app/components/layout/Header.tsx
@@ -8,7 +8,7 @@ import { useSession } from "next-auth/react";
 import Typography from "@mui/material/Typography";
 import Box from "@mui/material/Box";
 import { styled } from "@mui/material/styles";
-import Profile from "@/dashboard/app/components/auth/Profile";
+import Profile from "@bciers/components/navigation/Profile";
 // eslint-disable-next-line import/extensions
 import Logo from "@bciers/img/src/lib/BCID_CleanBC_rev_tagline_colour.svg";
