+++ bciers/apps/registration1/app/components/operations/OperationReview.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { useSession } from "next-auth/react";
 import Review from "@/app/components/button/Review";
 import { Status } from "@bciers/utils/enums";
