+++ bciers/apps/registration1/app/components/operations/OperationsForm.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import MultiStepFormBase from "@bciers/components/form/MultiStepFormBase";
 import Link from "next/link";
 import { useParams, useRouter } from "next/navigation";
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { useSession } from "next-auth/react";
 import { Status } from "@bciers/utils/enums";
 import { operationUiSchema } from "@/app/utils/jsonSchema/operations";
