+++ bciers/apps/registration1/app/components/routes/operators/Page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 import Loading from "@bciers/components/loading/SkeletonGrid";
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { GridRowsProp } from "@mui/x-data-grid";
 import buildQueryParams from "@bciers/utils/buildQueryParams";
 import { OperatorsSearchParams } from "@/app/components/userOperators/types";
