+++ bciers/apps/registration1/app/components/routes/select-operator/Page.tsx
@@ -4,7 +4,7 @@ import Loading from "@bciers/components/loading/SkeletonField";
 import { BC_GOV_LINKS_COLOR } from "@bciers/styles/colors";
 import Link from "next/link";
 import { auth } from "@/dashboard/auth";
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { OperatorStatus, UserOperatorStatus } from "@bciers/utils/enums";
 import getUserFullName from "@bciers/utils/getUserFullName";
 import SelectOperator from "@/app/components/userOperators/SelectOperator";
