+++ bciers/apps/registration1/app/components/userOperators/SelectOperatorForm.tsx
@@ -5,7 +5,7 @@ import { RJSFSchema } from "@rjsf/utils";
 import { useState } from "react";
 import { Alert } from "@mui/material";
 import { useRouter } from "next/navigation";
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import { SelectOperatorFormData } from "@/app/components/form/formDataTypes";
 import { selectOperatorUiSchema } from "@/app/utils/jsonSchema/selectOperator";
