+++ bciers/apps/reporting/src/app/(authenticated)/authenticated/operations/Operators.tsx
@@ -1,4 +1,4 @@
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 
 async function getOperators() {
   try {
