+++ bciers/apps/reporting/src/app/components/operations/Operations.tsx
@@ -1,4 +1,4 @@
-import { actionHandler } from "@bciers/utils/actions";
+import { actionHandler } from "@bciers/actions";
 import ReportingOperationDataGrid from "./ReportingOperationDataGrid";
 import { OperationRow } from "./types";
