+++ bciers/libs/shared/components/src/lib/navigation/Profile.tsx
@@ -1,7 +1,7 @@
 import { signOut, useSession } from "next-auth/react";
 import Button from "@mui/material/Button/Button";
 import Link from "@mui/material/Link";
-import { getEnvValue } from "@bciers/utils/actions";
+import { getEnvValue } from "@bciers/actions";
 import getUserFullName from "@bciers/utils/getUserFullName";
 
 export default function Profile() {
