+++ bciers/tsconfig.base.json
@@ -20,13 +20,13 @@
       "@/coam/*": ["apps/coam/*"],
       "@/dashboard/*": ["apps/dashboard/*"],
       "@/registration/*": ["apps/registration/*"],
-      "@bciers/actions/server": ["libs/shared/actions/src/server.ts"],
+      "@bciers/actions": ["libs/shared/actions/src/index.ts"],
       "@reporting/*": ["apps/reporting/*"],
       "@bciers/components": ["libs/shared/components/src/index.ts"],
       "@bciers/components/*": ["libs/shared/components/src/lib/*"],
       "@bciers/components/server": ["libs/shared/components/src/server.ts"],
       "@bciers/img/*": ["libs/shared/img/*"],
-      "@bciers/middlewares/server": ["libs/shared/middlewares/src/server.ts"],
+      "@bciers/middlewares": ["libs/shared/middlewares/src/index.ts"],
       "@bciers/styles": ["libs/shared/styles/src/index.ts"],
       "@bciers/styles/*": ["libs/shared/styles/src/lib/*"],
       "@bciers/testConfig/*": ["libs/shared/testConfig/src/*"],
